Add tools to prepare for STM-like scanning of additional positions
also refactor tools for combine imps and move some to imp_cluster_tools
to reuse the functions in there for the new STM-scanning feature
